Move more components into separate files
