+++ src/renderer/containers/Misc/Misc.tsx
@@ -1,15 +1,10 @@
 import * as React from "react";
 
-import styled, { css } from "styled-components";
-
-import { Tooltip } from "react-tippy";
-
 import { eventActionManager } from "@/actions";
 import { connectToOBS, setScene } from "@/lib/obs";
 import { ActionEvent } from "@/lib/realtime";
-import { getFilePath, getFolderPath, isDevelopment } from "@/lib/utils";
-import { Button, Icon, Input } from "semantic-ui-react";
-import { notify, openFileOrParentFolder } from "../../lib/utils";
+import { isDevelopment } from "@/lib/utils";
+import { notify } from "../../lib/utils";
 
 export const DevTools = () => {
     const handleClick = () => {
@@ -31,20 +26,6 @@ export const DevTools = () => {
     );
 };
 
-export const Labelled = (props: any) => {
-    const { onClick, children, ...rest } = props;
-    const pointerStyle = {
-        cursor: "pointer",
-    };
-    return (
-        <span style={onClick ? pointerStyle : undefined} onClick={onClick}>
-            <Tooltip arrow={true} duration={200} position="bottom" style={{ display: "inline-block" }} {...rest}>
-                {children}
-            </Tooltip>
-        </span>
-    );
-};
-
 export const CodeBlock: React.FC<{
     values: any
 }> = (props) => {
@@ -52,62 +33,4 @@ export const CodeBlock: React.FC<{
         return (<pre style={{overflowX: "auto"}}>{(JSON as any).stringify(props.values, 0, 2)}</pre>);
     }
     return null;
-};
-
-const NoMarginIcon = styled(Icon)`
-&&& {
-    margin: 0 !important;
-}
-`;
-
-interface FileInputProps extends Record<string, any> {
-    value: string;
-    onChange: (value: string) => void;
-    directory?: boolean;
-    fileTypeFilters?: Array<{name: string, extensions: string[]}>;
-    saveFile?: boolean;
-}
-
-export const FileInput: React.FC<FileInputProps> = props => {
-    const {value, directory, onChange, fileTypeFilters, saveFile} = props;
-    const [filesPath, setFilesPath] = React.useState<string>(value);
-    const selectFromFileSystem = async () => {
-        let p: string | null = null;
-        if (directory) {
-            // Handle directory selection
-            p = await getFolderPath();
-        } else {
-            // Handle file selection
-            let options: any;
-            if (fileTypeFilters) {
-                options = {
-                    filters: fileTypeFilters,
-                };
-            }
-            p = await getFilePath(options, saveFile);
-        }
-
-        if (p) {
-            setFilesPath(p);
-            onChange(p);
-        }
-    };
-    const actionLabel = saveFile ? "Save as" : "Choose";
-    // const IconWithTooltip = withTooltip(NoMarginIcon, { title: "Open location" });
-    return (
-        <Input
-            style={{ width: "100%" }}
-            label={
-                <Button onClick={() => openFileOrParentFolder(filesPath)}>
-                    <Labelled title="Open location">
-                        <NoMarginIcon name="folder open outline" />
-                    </Labelled>
-                </Button>
-            }
-            value={filesPath}
-            onChange={(_: any, { value }: any) => setFilesPath(value)}
-            onBlur={() => onChange(filesPath)}
-            action={<Button onClick={() => selectFromFileSystem().catch(console.error)}>{actionLabel}</Button>}
-        />
-    );
 };
