+++ src/renderer/containers/Settings/ComboFinder.tsx
@@ -8,7 +8,7 @@ import { loadFileInDolphin, notify, openComboInDolphin } from "@/lib/utils";
 import { Dispatch, iRootState } from "@/store";
 import { secondsToString } from "common/utils";
 import { InlineDropdown } from "../../components/InlineInputs";
-import { FileInput } from "../Misc/Misc";
+import { FileInput } from "@/components/FileInput";
 import { ProcessSection } from "../Misc/ProcessSection";
 import { RenameFiles } from "./RenameFiles";
